+++ public/models/User/User.js
@@ -2,14 +2,13 @@ export default class User {
   constructor() {
     this._id = null;
     this._login = null;
-    this._state = false;
   }
 
   get getUser() {
-    return this._state ? {
+    return {
       id: this._id,
       login: this._login
-    } : null;
+    };
   }
 
   get getLogin() {
@@ -20,18 +19,9 @@ export default class User {
     return this._id;
   }
 
-  get getState() {
-    return this._state;
-  }
-
   setUser(user = {}) {
     this.setLogin(user.login);
     this.setId(user.id);
-    this.setState(user.login !== null);
-  }
-
-  setState(state) {
-    this._state = state;
   }
 
   setLogin(login) {
